Debug traces
Some debug traces were added. The messages "frameGrabber Start Capturing" was removed. Also, in function CheckButtedJoint, the ResizeBitmap function was removed, it is not needed, since the size of easy ranger images and the size of the images used in the neural network matches (1009, 1450). In the case that we need to modify this size, we will need to create again the model with the new sizes